+++ lib/luna_park/errors/system.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'luna_park/errors/base'
+
+module LunaPark
+  module Errors
+    class System < Base; end
+  end
+end