+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2021-03-18
+Changed
+- Rename Interactors to UseCases
+- Rename Errors::Adaptive to Errors:Base 
+- Rename Errors::Processing to Errors::Business
+Added
+- Add class Errors::System
+- Add class UseCase::Service
+- Add class Extensions::HasError
+Removed
+- `on_error action:` instead of use Errors::Business and Errors::System directly
+
 ## [0.10.8] - 2021-02-09
 Added
 - Add sentry notifier